Criado IMPL para buscar os nos JCR, criado servlet para executar e retornar os nos

+++ core/src/main/java/com/vivopoc/aem/core/models/JsonModel.java
@@ -2,22 +2,22 @@ package com.vivopoc.aem.core.models;
 
 public class JsonModel {
 
-	int userId;
-	int id;
+	Long userId;
+	Long id;
 	String title;
 	boolean completed;
 	
 	
-	public int getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(int userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getTitle() {
